removed code duplication between the two parsers
